crm-5: refactoring permission to his trait
