+++ app/Models/HasPermissions.php
@@ -1,0 +1,46 @@
+<?php
+
+namespace App\Models;
+
+use Illuminate\Database\Eloquent\Collection;
+use Illuminate\Database\Eloquent\Relations\BelongsToMany;
+use Illuminate\Support\Facades\Cache;
+
+trait HasPermissions
+{
+    public function permissions(): BelongsToMany
+    {
+        return $this->belongsToMany(Permission::class);
+    }
+
+    public function givePermissionTo(string $key): void
+    {
+
+        $this->permissions()->firstOrCreate(compact('key'));
+
+        Cache::forget($this->getPermissionCacheKey());
+
+        Cache::rememberForever(
+            $this->getPermissionCacheKey(),
+            fn () => $this->permissions
+        );
+
+    }
+
+    public function hasPermissionTo(string $key): bool
+    {
+        /** @var Collection $permissions */
+        $permissions = Cache::get($this->getPermissionCacheKey(), $this->permissions);
+
+        //return $this->permissions()->where(compact('key'))->exists();
+
+        return $permissions
+                ->where('key', '=', $key)
+                ->isNotEmpty();
+    }
+
+    private function getPermissionCacheKey(): string
+    {
+        return "user::{$this->id}::permissions";
+    }
+}
