+++ app/Models/User.php
@@ -4,12 +4,9 @@ namespace App\Models;
 
 // use Illuminate\Contracts\Auth\MustVerifyEmail;
 
-use Illuminate\Database\Eloquent\Collection;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
-use Illuminate\Database\Eloquent\Relations\BelongsToMany;
 use Illuminate\Foundation\Auth\User as Authenticatable;
 use Illuminate\Notifications\Notifiable;
-use Illuminate\Support\Facades\Cache;
 use Laravel\Sanctum\HasApiTokens;
 
 class User extends Authenticatable
@@ -17,6 +14,7 @@ class User extends Authenticatable
     use HasApiTokens;
     use HasFactory;
     use Notifiable;
+    use HasPermissions;
 
     /**
      * The attributes that are mass assignable.
@@ -48,40 +46,4 @@ class User extends Authenticatable
         'email_verified_at' => 'datetime',
         'password'          => 'hashed',
     ];
-
-    public function permissions(): BelongsToMany
-    {
-        return $this->belongsToMany(Permission::class);
-    }
-
-    public function givePermissionTo(string $key): void
-    {
-
-        $this->permissions()->firstOrCreate(compact('key'));
-
-        Cache::forget($this->getPermissionCacheKey());
-
-        Cache::rememberForever(
-            $this->getPermissionCacheKey(),
-            fn () => $this->permissions
-        );
-
-    }
-
-    public function hasPermissionTo(string $key): bool
-    {
-        /** @var Collection $permissions */
-        $permissions = Cache::get($this->getPermissionCacheKey(), $this->permissions);
-
-        //return $this->permissions()->where(compact('key'))->exists();
-
-        return $permissions
-                ->where('key', '=', $key)
-                ->isNotEmpty();
-    }
-
-    private function getPermissionCacheKey(): string
-    {
-        return "user::{$this->id}::permissions";
-    }
 }
